producer/requirements.txt
praw
kafka-python-ng
tqdm
python-dotenv